+++ src/models/torrent_request_vote.rs
@@ -7,7 +7,7 @@ pub struct TorrentRequestVote {
     pub id: i64,
     pub torrent_request_id: i64,
     pub created_at: NaiveDateTime,
-    pub created_by_id: i32,
+    pub created_by_id: i64,
     pub bounty_upload: i64,
     pub bounty_bonus_points: i64,
 }
